lifecycle refactor
