+++ android/src/main/java/com/rnmapbox/rnmbx/components/mapview/RNMBXMapView.kt
@@ -13,10 +13,7 @@ import android.view.View.OnLayoutChangeListener
 import android.view.ViewGroup
 import android.widget.FrameLayout
 import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.LifecycleOwner
-import androidx.lifecycle.LifecycleRegistry
 import com.facebook.react.bridge.*
-import com.rnmapbox.rnmbx.utils.ViewCompat
 import com.mapbox.android.gestures.*
 import com.mapbox.bindgen.Value
 import com.mapbox.geojson.Feature
@@ -100,86 +97,11 @@ enum class MapGestureType {
     Move,Scale,Rotate,Fling,Shove
 }
 
-/***
- * Mapbox's MapView observers lifecycle events see MapboxLifecyclePluginImpl - (ON_START, ON_STOP, ON_DESTROY)
- * We need to emulate those.
- */
-interface RNMBXLifeCycleOwner : LifecycleOwner {
-    fun handleLifecycleEvent(event: Lifecycle.Event)
-}
-
 fun interface Cancelable {
     fun cancel()
 }
 
-class RNMBXLifeCycle {
-    private var lifecycleOwner : RNMBXLifeCycleOwner? = null
-
-    fun onAttachedToWindow(view: View) {
-        if (lifecycleOwner == null) {
-            lifecycleOwner = object : RNMBXLifeCycleOwner {
-                private lateinit var lifecycleRegistry: LifecycleRegistry
-                init {
-                    lifecycleRegistry = LifecycleRegistry(this)
-                    lifecycleRegistry.currentState = Lifecycle.State.CREATED
-                }
-
-                override fun handleLifecycleEvent(event: Lifecycle.Event) {
-                    try {
-                        lifecycleRegistry.handleLifecycleEvent(event)
-                    } catch (e: RuntimeException) {
-                        Log.e("RNMBXMapView", "handleLifecycleEvent, handleLifecycleEvent error: $e")
-                    }
-                }
-
-                override val lifecycle: Lifecycle
-                    get() = lifecycleRegistry
-
-            }
-            ViewCompat.setViewTreeLifecycleOwner(view, lifecycleOwner)
-        }
-        lifecycleOwner?.handleLifecycleEvent(Lifecycle.Event.ON_START)
-    }
-
-    fun onDetachedFromWindow() {
-        if (lifecycleOwner?.lifecycle?.currentState == Lifecycle.State.DESTROYED) {
-            return
-        }
-        lifecycleOwner?.handleLifecycleEvent(androidx.lifecycle.Lifecycle.Event.ON_STOP)
-    }
-
-    fun onDestroy() {
-        if (lifecycleOwner?.lifecycle?.currentState == Lifecycle.State.STARTED || lifecycleOwner?.lifecycle?.currentState == Lifecycle.State.RESUMED) {
-            lifecycleOwner?.handleLifecycleEvent(androidx.lifecycle.Lifecycle.Event.ON_STOP)
-        }
-        if (lifecycleOwner?.lifecycle?.currentState != Lifecycle.State.DESTROYED) {
-            lifecycleOwner?.handleLifecycleEvent(androidx.lifecycle.Lifecycle.Event.ON_DESTROY)
-        }
-    }
-
-    fun getState() : Lifecycle.State {
-        return lifecycleOwner?.lifecycle?.currentState ?: Lifecycle.State.INITIALIZED;
-    }
-
-    var attachedToWindowWaiters : MutableList<()-> Unit> = mutableListOf()
-
-    fun callIfAttachedToWindow(callback: () -> Unit) : com.rnmapbox.rnmbx.components.mapview.Cancelable {
-        if (getState() == Lifecycle.State.STARTED) {
-            callback()
-            return com.rnmapbox.rnmbx.components.mapview.Cancelable {}
-        } else {
-            attachedToWindowWaiters.add(callback)
-            return com.rnmapbox.rnmbx.components.mapview.Cancelable {
-                attachedToWindowWaiters.removeIf21 { it === callback }
-            }
-        }
-    }
-
-    fun afterAttachFromLooper() {
-        attachedToWindowWaiters.forEach { it() }
-        attachedToWindowWaiters.clear()
-    }
-}
+// RNMBXLifeCycle is now provided by lifecycle-compat layer
 
 data class FeatureEntry(val feature: AbstractMapFeature?, val view: View?, var addedToMap: Boolean = false) {
